+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 0.1.1 (2010-12-10)
 ------------------
 
+- Added an examples folder with a single example showing a script to send
+  an email summarizing stories across projects to multiple team members.
+
 - Provide both ``get()`` and ``get_etree()``.  The ``get()`` method
   returns the response and content of the request directly from httplib2.
   The ``get_etree()`` will return an ElementTree object of the content.